apps/policy-gateway/requirements-dev.txt
fastapi>=0.95.0
pydantic>=1.10.0
uvicorn>=0.22.0
pytest>=7.0.0
requests>=2.28.0
httpx>=0.24.0
ruff>=0.12.0
# Optional: add testing helpers
pytest-cov>=4.0.0
PyYAML>=6.0
litellm>=0.1.0
pytest-asyncio>=0.21.0

